Improvements to how API runs in conjunction with localization.
There was actually a bug previously where text domains were getting
registered after the API was running, and so the result was that none
of the text strings established with the API system like the
construction of the Theme Options (the most notable example) were
actually available for localization.

So, now I've added three do_action's to themeblvd.php after everything
has been hooked.

(1) The first fires a hook not being used by the framework called
"themeblvd_intervene" which allows for a Child theme to potentially
hook in between all framework hooked action and the localization/API
running (I honestly can't think of why anyone would even do this as of
now, but it is there!).

(2) The second do_action added has the function hooked to it that
registers the text domains for localization. I wanted this added as an
action, so someone could easily unhook and re-locate the localization
files to their Child theme if desired.

(3) The third do_action then runs the API system. It is crucial that
this happens in this file so all Child theme API functions work
correctly. In other words, this CANNOT be hooked to "after_setup_theme"
-- However, instead of before how we simply ran the
themeblvd_api_init() function, now it's hooked to this action. This
means someone could potentially override this utilizing the
themeblvd_intervene hook, but as stated above, I honestly can't think
of a reason why someone would be doing this; although who knows what
kind of cool stuff we'll doing with child themes and plugins in the
future though!