Improved logo image handling for retina
In order to to allow for a swap of the 2x logo image on retina, we need
to know the width of the original image. Originally I was doing this
with jQuery, but this can be inconsistent.

So, I've adjusted this to work by detecting the image size of the
original size image with PHP when the logo is displayed. And so we're
not having to read the image file every page load with getimagesize(),
I've stored this in a transient. The transient is deleted every time
the theme options are saved, as
`delete_transient('themeblvd_logo_size')` was added to the logo
option's sanitization.