Improvements to TGM plugin class.
Two major things were altered in the class:

1) I got rid of add_settings_error and settings_error functions. Since
the class's notices function is already hooked to admin_notices, I
simply outputted the notices instead. Registering them as
settings_errors causes conflicts with too many other things.

2) Adjusted how the class checks to show the nag or not. -- Before the
class was setup to check for a meta value attached to the user that was
the same no matter what. So, this means if the the user switched to
another theme with this class, they would never even see the prompt. I
adjusted this to use the theme's identifier so changing themes will
cause the prompt to show.