Changed my mind. Put image sizes back to where they were.
Downside is we can't display image sizes to the user with things setup
at the API level, like options of builder elements, for example.

Upside is that devs can still filter image sizes in the same way after
the update. If we kept this change, everyone using the filter would
have to to know to move it before the framework.