Added multiple ways of setting categories in themeblvd_get_posts_args.
The main prupose of this was to add the ability from the post list/post
grid shortcodes in the Theme Blvd shortcodes plugin to easier pull
categories. This actually saves resources, and a DB query, as well.

Also implemented in the themeblvd_posts() function to allow for you to
force a category string. In most cases this is getting called from a
custom layout, so really this makes no difference. But if someone were
to call themeblvd_posts() directly, they could feed in "cat" or
"category_name" in the $args, and it will actually save some resources
as mentioned in the first paragraph.