Moved analytics to <head>
This is sort of interesting topic. Practically, with Google
asynchronous code, it doesn't really matter where it goes in the site.

When Google gives you the snippet, they tell you to put it "immediately
after the opening <body> tag" -- but this directly contradicts what it
explain in detail from their documentation.

I read an interesting opinion somewhere that points out that it's
possible a user will go to your website and leave before the page has
fully loaded. And so it's possible Google tracking may not have kicked
in yet. By putting the snippet closer to the top of the document, it
makes this less likely.

Also... We're putting it in the exact same place as the famous
Analytics plugin by Yoast. Can't go wrong with that approach!