Incorporated class "tb-options-js"
As we've expanded on the options framework, we've ran into different
scenarios where there are multiple option sets on a single page. For
example, maybe options are incorporated into two separate meta boxes
when editing a page.

This doesn't work with the javascript being attached to element with ID
"optionsframework" because jQuery isn't expecting there to be multiple
elements with one ID.