Header info option for su/exec bases (see desc)
It used to be the the Super User base displayed header info as a bar
across the top of the header, while the Executive base displayed header
info within the header’s content next to the logo. Now, both bases have
an option, allowing the user to pick which one they want to do in
either base.