Better handling of empty theme options (see desc)
Previously, the system was setup to fill in option defaults if the
ENTIRE options array was empty. However, if the user had theme options
saved, then there was a new option added, that value would be blank.
This fixes that, and allows for more smoother updates, when the theme
has a new option but user doesn’t know to go to theme options and
re-save.