Simplify header styling transition (see desc). #280
Previously in framework 2.6, we added a consistent mobile header style throughout all theme bases. In this commit, this was improved to use Flexbox CSS (to help with #316). Also this commit applies this consistent mobile header from 991px and below (instead of just 767px and below). This way, we don't have a "limbo" stage of the header between desktop and mobile.